+++ scripts/ammunition-system/actions/reload-magazine.js
@@ -1,4 +1,7 @@
 import { handleReload } from "../../feats/crossbow-feats.js";
+import { Weapon } from "../../types/pf2e-ranged-combat/weapon.js";
+import { PF2eActor } from "../../types/pf2e/actor.js";
+import { PF2eToken } from "../../types/pf2e/token.js";
 import { getControlledActorAndToken, getEffectFromActor, getFlag, getItem, postInChat, setEffectTarget, showWarning, Updates, useAdvancedAmmunitionSystem } from "../../utils/utils.js";
 import { getWeapon } from "../../utils/weapon-utils.js";
 import { MAGAZINE_LOADED_EFFECT_ID, RELOAD_MAGAZINE_IMG } from "../constants.js";
@@ -45,6 +48,16 @@ export async function reloadMagazine() {
         return;
     }
 
+    performReloadMagazine(actor, token, weapon);
+}
+
+/**
+ * @param {PF2eActor} actor 
+ * @param {PF2eToken} token 
+ * @param {Weapon} weapon 
+ * @returns 
+ */
+export async function performReloadMagazine(actor, token, weapon) {
     const updates = new Updates(actor);
 
     // If we have no ammunition selected, or we have none left in the stack, we can't reload