+++ scripts/config.js
@@ -5,10 +5,11 @@ import { consolidateRepeatingWeaponAmmunition } from "./ammunition-system/action
 import { fireBothBarrels } from "./ammunition-system/actions/fire-both-barrels.js";
 import { nextChamber } from "./ammunition-system/actions/next-chamber.js";
 import { reloadMagazine } from "./ammunition-system/actions/reload-magazine.js";
-import { reload, reloadNPCs, fullyReload } from "./ammunition-system/actions/reload.js";
+import { fullyReload, reload, reloadNPCs } from "./ammunition-system/actions/reload.js";
 import { switchAmmunition } from "./ammunition-system/actions/switch-ammunition.js";
 import { unload } from "./ammunition-system/actions/unload.js";
 import { huntPrey } from "./hunt-prey/hunt-prey.js";
+import { initialiseLibWrapperHooks } from "./libwrapper-hooks.js";
 import { npcWeaponConfiguration } from "./npc-weapon-system/npc-weapon-system.js";
 import { runMigrations } from "./utils/migrations/migration.js";
 
@@ -140,6 +141,8 @@ Hooks.on(
                 config: false
             }
         );
+
+        initialiseLibWrapperHooks();
 
         game.pf2eRangedCombat = {
             reload,